# docs/openapi/issuer-service.yaml
openapi: 3.0.3
info:
  title: CUNCTA Issuer Service
  version: 1.0.0-testnet
servers:
  - url: http://localhost:3002
paths:
...
  /token:
    post:
      summary: Token (pre-authorized code)
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
      responses:
        "200":
          description: Access token
  /credential:
    post:
      summary: Issue credential (OID4VCI-style)
      security:
        - accessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Issued credential
  /v1/issue:
    post:
      summary: Issue credential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Issued credential
  /v1/internal/issue:
    post:
      summary: Issue credential (service JWT required)
      security:
        - serviceJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Issued credential
  /v1/aura/claim:
    post:
      summary: Claim aura credential (service JWT required)
      security:
        - serviceJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Aura claim response
...
components:
  securitySchemes:
    serviceJwt:
      type: http
      scheme: bearer
    accessToken:
      type: http
      scheme: bearer
    dsrToken:
      type: http
      scheme: bearer